fix: handle dlc misclassification
A "DOWNLOAD REQUIRED" DLC would get incorrectly classified as "PRE-INSTALLED" if the Depot ID for that DLC was not the same as its App ID. For example, `Assassin's Creed Shadows (3159330)` has the DLC `Assassin's Creed Shadows - Claws of Awaji (3362550)`, but the Depot ID for it is `3159339`. `Cyberpunk (1091500)`, on the other hand, has the DLC `Cyberpunk 2077 REDmod (2060310)`, with the same Depot ID of `2060310`.